+++ common/src/main/java/com/st0x0ef/stellaris/common/items/armors/AbstractSpaceArmor.java
@@ -38,12 +38,14 @@ public abstract class AbstractSpaceArmor extends CustomArmorItem {
 
         @Override
         public @NotNull UniversalFluidItemStorage getFluidTank(@NotNull ItemStack stack) {
-
-            return new ItemFluidStorage(DataComponentsRegistry.FLUID_LIST.get(), stack, 1, 3000) {
-
+            return new ItemFluidStorage(DataComponentsRegistry.FLUID_LIST.get(), stack, 2, 3000) {
                 @Override
                 public boolean isFluidValid(int tank, FluidStack stack) {
-                    return stack.getFluid().isSame(FluidRegistry.OXYGEN_STILL.get());
+                    return switch (tank) {
+                        case 0 -> stack.getFluid().isSame(FluidRegistry.OXYGEN_STILL.get());
+                        case 1 -> stack.getFluid().isSame(FluidRegistry.DIESEL_STILL.get());
+                        default -> false;
+                    };
                 }
             };
         }
@@ -58,7 +60,7 @@ public abstract class AbstractSpaceArmor extends CustomArmorItem {
         @Override
         public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag) {
             super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
-            tooltipComponents.add(Component.translatable("jetsuit.stellaris.fuel", getFluidTank(stack).getFluidInTank(1).getAmount()));
+            tooltipComponents.add(Component.translatable("jetsuit.stellaris.diesel", getFluidTank(stack).getFluidInTank(1).getAmount()));
 
         }
 
@@ -70,7 +72,7 @@ public abstract class AbstractSpaceArmor extends CustomArmorItem {
                 public boolean isFluidValid(int tank, FluidStack stack) {
                     return switch (tank) {
                         case 0 -> stack.getFluid().isSame(FluidRegistry.OXYGEN_STILL.get());
-                        case 1 -> stack.getFluid().isSame(FluidRegistry.FUEL_STILL.get());
+                        case 1 -> stack.getFluid().isSame(FluidRegistry.DIESEL_STILL.get());
                         default -> false;
                     };
                 }