+++ changelog.md
@@ -2,7 +2,11 @@
 
 ## Additions
 - Add diesel
-- Add diesel diesel generator
+- Add diesel generator
 
 ## Changes
 - Rover now need diesel instead of fuel
+- Jet suit and space suit now need diesel instead of fuel
+
+## Fixes
+- Space suit crash when trying to put fuel in it